Move config to its own package
