+++ exporter/datadogexporter/factory.go
@@ -20,14 +20,13 @@ import (
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/config/confignet"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/config"
 )
 
 const (
 	// typeStr is the type of the exporter
 	typeStr = "datadog"
-
-	// DefaultSite is the default site of the Datadog intake to send data to
-	DefaultSite = "datadoghq.com"
 )
 
 // NewFactory creates a Datadog exporter factory
@@ -41,31 +40,31 @@ func NewFactory() component.ExporterFactory {
 
 // createDefaultConfig creates the default exporter configuration
 func createDefaultConfig() configmodels.Exporter {
-	return &Config{
+	return &config.Config{
 		ExporterSettings: configmodels.ExporterSettings{
 			TypeVal: configmodels.Type(typeStr),
 			NameVal: typeStr,
 		},
 
-		TagsConfig: TagsConfig{
+		TagsConfig: config.TagsConfig{
 			Hostname: "${DD_HOST}",
 			Env:      "${DD_ENV}",
 			Service:  "${DD_SERVICE}",
 			Version:  "${DD_VERSION}",
 		},
 
-		API: APIConfig{
+		API: config.APIConfig{
 			Key:  "", // must be set if using API
-			Site: DefaultSite,
+			Site: config.DefaultSite,
 		},
 
-		Metrics: MetricsConfig{
+		Metrics: config.MetricsConfig{
 			TCPAddr: confignet.TCPAddr{
 				Endpoint: "", // set during config sanitization
 			},
 		},
 
-		Traces: TracesConfig{
+		Traces: config.TracesConfig{
 			SampleRate: 1,
 			TCPAddr: confignet.TCPAddr{
 				Endpoint: "", // set during config sanitization
@@ -81,7 +80,7 @@ func createMetricsExporter(
 	c configmodels.Exporter,
 ) (component.MetricsExporter, error) {
 
-	cfg := c.(*Config)
+	cfg := c.(*config.Config)
 
 	params.Logger.Info("sanitizing Datadog metrics exporter configuration")
 	if err := cfg.Sanitize(); err != nil {
