+++ exporter/datadogexporter/factory_test.go
@@ -24,8 +24,11 @@ import (
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config/configcheck"
 	"go.opentelemetry.io/collector/config/configmodels"
+	"go.opentelemetry.io/collector/config/confignet"
 	"go.opentelemetry.io/collector/config/configtest"
 	"go.uber.org/zap"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/config"
 )
 
 // Test that the factory creates the default configuration
@@ -33,27 +36,83 @@ func TestCreateDefaultConfig(t *testing.T) {
 	factory := NewFactory()
 	cfg := factory.CreateDefaultConfig()
 
-	assert.Equal(t, &Config{
+	assert.Equal(t, &config.Config{
 		ExporterSettings: configmodels.ExporterSettings{
 			TypeVal: configmodels.Type(typeStr),
 			NameVal: typeStr,
 		},
 
 		// These are filled when loading using the helper methods
-		TagsConfig: TagsConfig{
+		TagsConfig: config.TagsConfig{
 			Hostname: "${DD_HOST}",
 			Env:      "${DD_ENV}",
 			Service:  "${DD_SERVICE}",
 			Version:  "${DD_VERSION}",
 		},
 
-		API: APIConfig{Site: "datadoghq.com"},
-		Traces: TracesConfig{
+		API: config.APIConfig{Site: "datadoghq.com"},
+		Traces: config.TracesConfig{
 			SampleRate: 1,
 		},
 	}, cfg, "failed to create default config")
 
 	assert.NoError(t, configcheck.ValidateConfig(cfg))
+}
+
+// TestLoadConfig tests that the configuration is loaded correctly
+func TestLoadConfig(t *testing.T) {
+	factories, err := componenttest.ExampleComponents()
+	assert.NoError(t, err)
+
+	factory := NewFactory()
+	factories.Exporters[typeStr] = factory
+	cfg, err := configtest.LoadConfigFile(t, path.Join(".", "testdata", "config.yaml"), factories)
+
+	require.NoError(t, err)
+	require.NotNil(t, cfg)
+
+	apiConfig := cfg.Exporters["datadog/api"].(*config.Config)
+	err = apiConfig.Sanitize()
+
+	require.NoError(t, err)
+	assert.Equal(t, &config.Config{
+		ExporterSettings: configmodels.ExporterSettings{
+			NameVal: "datadog/api",
+			TypeVal: typeStr,
+		},
+
+		TagsConfig: config.TagsConfig{
+			Hostname: "customhostname",
+			Env:      "prod",
+			Service:  "myservice",
+			Version:  "myversion",
+			Tags:     []string{"example:tag"},
+		},
+
+		API: config.APIConfig{
+			Key:  "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+			Site: "datadoghq.eu",
+		},
+
+		Metrics: config.MetricsConfig{
+			Namespace: "opentelemetry.",
+			TCPAddr: confignet.TCPAddr{
+				Endpoint: "https://api.datadoghq.eu",
+			},
+		},
+
+		Traces: config.TracesConfig{
+			SampleRate: 1,
+			TCPAddr: confignet.TCPAddr{
+				Endpoint: "https://trace.agent.datadoghq.eu",
+			},
+		},
+	}, apiConfig)
+
+	invalidConfig2 := cfg.Exporters["datadog/invalid"].(*config.Config)
+	err = invalidConfig2.Sanitize()
+	require.Error(t, err)
+
 }
 
 func TestCreateAPIMetricsExporter(t *testing.T) {
